added battery mode: dynamically change key touch threshold
when the kb is powered on, it will use the default configured threshold
in `TOUCH_THRESHOLD` or if defined individually for a key.

But when the laptop is on battery, sometimes the values when a key is
touched won't go below the default threshold.

But setting the default threshold too high to accomodate for the battery
mode result in neighbouring keys to be detected as
"touched", since then the value read from the touch pin lies between
0-7, whereas on battery the
values vary alot in the range 15-30

To dynamically detect when to use the secondary threshold tailored for
laptop being on battery, the user has to touch one of the keys when
keyboard is being powered on. If the laptop is on battery, the minimum
value when pin is touched will be b/w 15-30. This is the indicator of
laptop being on battery.